Refactor redux components into separate files
