+++ client-v2/src/redux/configure-store.js
@@ -1,0 +1,28 @@
+import { createStore, applyMiddleware, compose } from 'redux'
+import thunk from 'redux-thunk'
+import { persistStore, persistReducer } from 'redux-persist'
+import storage from 'redux-persist/lib/storage'
+
+import rootReducer from './reducers'
+
+const persistConfig = {
+  key: 'root',
+  storage,
+}
+
+const persistedReducer = persistReducer(persistConfig, rootReducer)
+
+export default () => {
+  const composeEnhancers =
+    (typeof window !== 'undefined' &&
+      window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
+    compose
+
+  const store = createStore(
+    persistedReducer,
+    composeEnhancers(applyMiddleware(thunk)),
+  )
+
+  const persistor = persistStore(store)
+  return { store, persistor }
+}
